Move graph driver registration out of the daemon package.
