+++ daemon/graphdriver/register/register_aufs.go
@@ -1,6 +1,6 @@
 // +build !exclude_graphdriver_aufs,linux
 
-package daemon
+package register
 
 import (
 	// register the aufs graphdriver
