+++ daemon/graphdriver/register/register_btrfs.go
@@ -1,6 +1,6 @@
 // +build !exclude_graphdriver_btrfs,linux
 
-package daemon
+package register
 
 import (
 	// register the btrfs graphdriver
