+++ daemon/graphdriver/register/register_vfs.go
@@ -1,0 +1,6 @@
+package register
+
+import (
+	// register vfs
+	_ "github.com/docker/docker/daemon/graphdriver/vfs"
+)
