+++ daemon/graphdriver/register/register_windows.go
@@ -1,0 +1,6 @@
+package register
+
+import (
+	// register the windows graph driver
+	_ "github.com/docker/docker/daemon/graphdriver/windows"
+)
